Description
Fix imprecise transaction cases, adjust test:

Entity, DTO, Controller, Services, etc : Change BigDecimal syntax, function.
Test: Implement TestContainer

+++ src/main/java/lumi/insert/app/core/repository/projection/ProductOutOfStock.java
@@ -1,8 +1,10 @@
 package lumi.insert.app.core.repository.projection;
 
+import java.math.BigDecimal;
+
 public record ProductOutOfStock(
     Long id,
     String name,
-    Long stockQuantity,
-    Long stockMinimum
+    BigDecimal stockQuantity,
+    BigDecimal stockMinimum
 ) {}
