+++ src/main/java/lumi/insert/app/dto/response/SupplyItemResponse.java
@@ -1,5 +1,6 @@
 package lumi.insert.app.dto.response;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 import io.swagger.v3.oas.annotations.media.Schema;
 
@@ -13,10 +14,10 @@ public record SupplyItemResponse(
     ProductName product, 
     
     @Schema(description = "Unit cost price charged by the supplier for this specific delivery", example = "12500")
-    Long price, 
+    BigDecimal price, 
     
     @Schema(description = "Quantity of the product received from the supplier", example = "100")
-    Long quantity, 
+    BigDecimal quantity, 
     
     @Schema(description = "Additional notes or specifications for this batch (e.g., Expiry Date, Batch No)", example = "Batch #A2024 - Exp: 2026-12")
     String description
